Fix bug: music type can be displayed at table.

+++ src/Music.java
@@ -8,14 +8,13 @@ public class Music extends Product{
 
     private String singer;
     private String genre;
-//    private int year;
     private int numOfSongs;
 
-    public Music(String name, float price, int quantity, ProductType type, int year){
+    public Music(String name, double price, int quantity, ProductType type, int year){
         super(name, price, quantity, type,year);
     }
 
-    public Music(ProductType type, String name, float price, int year, String genre, String singer, int quantity, int numOfSongs){
+    public Music(ProductType type, String name, double price, int year, String genre, String singer, int quantity, int numOfSongs){
         super(name, price, quantity, type,year);
         this.singer = singer;
         this.genre = genre;
